extract drawing mask functions into separate module; perform slice interp per mask color
