+++ packages/niivue/src/niivue/index.ts
@@ -114,6 +114,7 @@ import {
   SyncOpts
 } from '../types.js'
 import { toNiivueObject3D } from '../nvimage/RenderingUtils.js'
+import { findBoundarySlices, interpolateMaskSlices } from '../drawing/drawing.js'
 import {
   clamp,
   decodeRLE,
@@ -15088,89 +15089,19 @@ export class Niivue {
       return null
     }
 
-    const dims = this.back.dims
-    const [dimX, dimY, dimZ] = [dims[1], dims[2], dims[3]]
-
-    let axisSize: number
-    if (sliceType === SLICE_TYPE.AXIAL) {
-      axisSize = dimZ
-    } else if (sliceType === SLICE_TYPE.CORONAL) {
-      axisSize = dimY
-    } else if (sliceType === SLICE_TYPE.SAGITTAL) {
-      axisSize = dimX
-    } else {
-      return null
-    }
-
-    let firstSliceWithData = -1
-    let lastSliceWithData = -1
-
-    for (let slice = 0; slice < axisSize; slice++) {
-      let hasData = false
-
-      if (sliceType === SLICE_TYPE.AXIAL) {
-        // Check axial slice (XY plane at Z=slice)
-        const offset = slice * dimX * dimY
-        for (let i = 0; i < dimX * dimY; i++) {
-          if (this.drawBitmap[offset + i] > 0) {
-            hasData = true
-            break
-          }
-        }
-      } else if (sliceType === SLICE_TYPE.CORONAL) {
-        // Check coronal slice (XZ plane at Y=slice)
-        for (let z = 0; z < dimZ; z++) {
-          for (let x = 0; x < dimX; x++) {
-            const idx = x + slice * dimX + z * dimX * dimY
-            if (this.drawBitmap[idx] > 0) {
-              hasData = true
-              break
-            }
-          }
-          if (hasData) {
-            break
-          }
-        }
-      } else if (sliceType === SLICE_TYPE.SAGITTAL) {
-        // Check sagittal slice (YZ plane at X=slice)
-        for (let z = 0; z < dimZ; z++) {
-          for (let y = 0; y < dimY; y++) {
-            const idx = slice + y * dimX + z * dimX * dimY
-            if (this.drawBitmap[idx] > 0) {
-              hasData = true
-              break
-            }
-          }
-          if (hasData) {
-            break
-          }
-        }
-      }
-
-      if (hasData) {
-        if (firstSliceWithData === -1) {
-          firstSliceWithData = slice
-        }
-        lastSliceWithData = slice
-      }
-    }
-
-    if (firstSliceWithData === -1 || lastSliceWithData === -1) {
-      return null
-    }
-
-    return { first: firstSliceWithData, last: lastSliceWithData }
+    const dims = { dimX: this.back.dims[1], dimY: this.back.dims[2], dimZ: this.back.dims[3] }
+    return findBoundarySlices(sliceType, this.drawBitmap, dims)
   }
 
   /**
    * Interpolate between mask slices using geometric or intensity-guided methods
-   * @param sliceIndexLow - Lower slice index
-   * @param sliceIndexHigh - Higher slice index
+   * @param sliceIndexLow - Lower slice index (optional, will auto-detect if not provided)
+   * @param sliceIndexHigh - Higher slice index (optional, will auto-detect if not provided)
    * @param options - Interpolation options
    */
-  interpolateBetweenMaskSlices(
-    sliceIndexLow: number,
-    sliceIndexHigh: number,
+  interpolateMaskSlices(
+    sliceIndexLow?: number,
+    sliceIndexHigh?: number,
     options: {
       intensityWeight?: number
       binaryThreshold?: number
@@ -15184,320 +15115,12 @@ export class Niivue {
       throw new Error('Background image and drawing bitmap must be loaded')
     }
 
-    if (sliceIndexLow >= sliceIndexHigh) {
-      throw new Error('Low slice index must be less than high slice index')
-    }
-
-    const dims = this.back.dims
-    const [dimX, dimY, dimZ] = [dims[1], dims[2], dims[3]]
-
-    // Determine slice type (default to axial)
-    const sliceType = options.sliceType ?? SLICE_TYPE.AXIAL
-    // Get dimensions based on slice type
-    let sliceWidth: number, sliceHeight: number, maxSliceIndex: number
-    if (sliceType === SLICE_TYPE.AXIAL) {
-      sliceWidth = dimX
-      sliceHeight = dimY
-      maxSliceIndex = dimZ - 1
-    } else if (sliceType === SLICE_TYPE.CORONAL) {
-      sliceWidth = dimX
-      sliceHeight = dimZ
-      maxSliceIndex = dimY - 1
-    } else if (sliceType === SLICE_TYPE.SAGITTAL) {
-      sliceWidth = dimY
-      sliceHeight = dimZ
-      maxSliceIndex = dimX - 1
-    } else {
-      throw new Error('Invalid slice type. Must be AXIAL, CORONAL, or SAGITTAL')
-    }
-
-    if (sliceIndexLow < 0 || sliceIndexHigh > maxSliceIndex) {
-      throw new Error(`Slice indices out of bounds [0, ${maxSliceIndex}]`)
-    }
-
-    // Set default options
-    const opts = {
-      intensityWeight: options.intensityWeight ?? 0.7,
-      binaryThreshold: options.binaryThreshold ?? 0.375,
-      intensitySigma: options.intensitySigma ?? 0.1,
-      applySmoothingToSlices: options.applySmoothingToSlices ?? true,
-      useIntensityGuided: options.useIntensityGuided ?? true
-    }
-
-    // Extract boundary slices
-    const sliceLow = this.extractSlice(sliceIndexLow, sliceType)
-    const sliceHigh = this.extractSlice(sliceIndexHigh, sliceType)
-
-    // Skip interpolation if either boundary slice is empty
-    if (!this.hasSliceVariability(sliceLow) || !this.hasSliceVariability(sliceHigh)) {
-      return
-    }
-
-    // Apply smoothing if enabled
-    if (opts.applySmoothingToSlices) {
-      this.smoothSlice(sliceLow, sliceWidth, sliceHeight)
-      this.smoothSlice(sliceHigh, sliceWidth, sliceHeight)
-    }
-
-    // Interpolate between boundary slices
-    for (let z = sliceIndexLow + 1; z < sliceIndexHigh; z++) {
-      const interpolatedSlice = new Float32Array(sliceWidth * sliceHeight)
-
-      if (opts.useIntensityGuided && this.back.img) {
-        // Intensity-guided interpolation
-        this.performIntensityGuidedInterpolation(
-          sliceLow,
-          sliceHigh,
-          z,
-          sliceIndexLow,
-          sliceIndexHigh,
-          interpolatedSlice,
-          opts,
-          sliceType
-        )
-      } else {
-        // Geometric interpolation
-        this.performGeometricInterpolation(sliceLow, sliceHigh, z, sliceIndexLow, sliceIndexHigh, interpolatedSlice)
-      }
-
-      // Insert interpolated slice back into drawing bitmap
-      this.insertSlice(interpolatedSlice, z, sliceType, opts.binaryThreshold)
-    }
-
-    // Update the drawing texture
-    this.refreshDrawing(true)
-  }
-
-  private extractSlice(sliceIndex: number, sliceType: SLICE_TYPE): Float32Array {
-    if (!this.back || !this.back.dims || !this.drawBitmap) {
-      throw new Error('Background image and drawing bitmap required')
-    }
-
-    const dims = this.back.dims
-    const [dimX, dimY, dimZ] = [dims[1], dims[2], dims[3]]
-
-    let sliceData: Float32Array
-    if (sliceType === SLICE_TYPE.AXIAL) {
-      // Extract axial slice (XY plane at Z=sliceIndex)
-      sliceData = new Float32Array(dimX * dimY)
-      const offset = sliceIndex * dimX * dimY
-      for (let i = 0; i < dimX * dimY; i++) {
-        sliceData[i] = this.drawBitmap[offset + i]
-      }
-    } else if (sliceType === SLICE_TYPE.CORONAL) {
-      // Extract coronal slice (XZ plane at Y=sliceIndex)
-      sliceData = new Float32Array(dimX * dimZ)
-      for (let z = 0; z < dimZ; z++) {
-        for (let x = 0; x < dimX; x++) {
-          const srcIdx = x + sliceIndex * dimX + z * dimX * dimY
-          const dstIdx = x + z * dimX
-          sliceData[dstIdx] = this.drawBitmap[srcIdx]
-        }
-      }
-    } else if (sliceType === SLICE_TYPE.SAGITTAL) {
-      // Extract sagittal slice (YZ plane at X=sliceIndex)
-      sliceData = new Float32Array(dimY * dimZ)
-      for (let z = 0; z < dimZ; z++) {
-        for (let y = 0; y < dimY; y++) {
-          const srcIdx = sliceIndex + y * dimX + z * dimX * dimY
-          const dstIdx = y + z * dimY
-          sliceData[dstIdx] = this.drawBitmap[srcIdx]
-        }
-      }
-    } else {
-      throw new Error('Invalid slice type')
-    }
-
-    return sliceData
-  }
-
-  private extractIntensitySlice(sliceIndex: number, sliceType: SLICE_TYPE): Float32Array {
-    if (!this.back || !this.back.dims || !this.back.img) {
-      throw new Error('Background image required for intensity data')
-    }
-
-    const dims = this.back.dims
-    const [dimX, dimY, dimZ] = [dims[1], dims[2], dims[3]]
+    const dims = { dimX: this.back.dims[1], dimY: this.back.dims[2], dimZ: this.back.dims[3] }
+    const imageData = this.back.img
     const maxVal = this.back.global_max
 
-    let sliceData: Float32Array
-    if (sliceType === SLICE_TYPE.AXIAL) {
-      // Extract axial slice (XY plane at Z=sliceIndex)
-      sliceData = new Float32Array(dimX * dimY)
-      const offset = sliceIndex * dimX * dimY
-      for (let i = 0; i < dimX * dimY; i++) {
-        sliceData[i] = this.back.img[offset + i] / maxVal
-      }
-    } else if (sliceType === SLICE_TYPE.CORONAL) {
-      // Extract coronal slice (XZ plane at Y=sliceIndex)
-      sliceData = new Float32Array(dimX * dimZ)
-      for (let z = 0; z < dimZ; z++) {
-        for (let x = 0; x < dimX; x++) {
-          const srcIdx = x + sliceIndex * dimX + z * dimX * dimY
-          const dstIdx = x + z * dimX
-          sliceData[dstIdx] = this.back.img[srcIdx] / maxVal
-        }
-      }
-    } else if (sliceType === SLICE_TYPE.SAGITTAL) {
-      // Extract sagittal slice (YZ plane at X=sliceIndex)
-      sliceData = new Float32Array(dimY * dimZ)
-      for (let z = 0; z < dimZ; z++) {
-        for (let y = 0; y < dimY; y++) {
-          const srcIdx = sliceIndex + y * dimX + z * dimX * dimY
-          const dstIdx = y + z * dimY
-          sliceData[dstIdx] = this.back.img[srcIdx] / maxVal
-        }
-      }
-    } else {
-      throw new Error('Invalid slice type')
-    }
-
-    return sliceData
-  }
-
-  private insertSlice(slice: Float32Array, sliceIndex: number, sliceType: SLICE_TYPE, binaryThreshold: number): void {
-    if (!this.drawBitmap || !this.back || !this.back.dims) {
-      throw new Error('Drawing bitmap and background required')
-    }
-
-    const dims = this.back.dims
-    const [dimX, dimY, dimZ] = [dims[1], dims[2], dims[3]]
-
-    if (sliceType === SLICE_TYPE.AXIAL) {
-      // Insert axial slice (XY plane at Z=sliceIndex)
-      const offset = sliceIndex * dimX * dimY
-      for (let i = 0; i < slice.length; i++) {
-        this.drawBitmap[offset + i] = slice[i] >= binaryThreshold ? this.opts.penValue : 0
-      }
-    } else if (sliceType === SLICE_TYPE.CORONAL) {
-      // Insert coronal slice (XZ plane at Y=sliceIndex)
-      for (let z = 0; z < dimZ; z++) {
-        for (let x = 0; x < dimX; x++) {
-          const srcIdx = x + z * dimX
-          const dstIdx = x + sliceIndex * dimX + z * dimX * dimY
-          this.drawBitmap[dstIdx] = slice[srcIdx] >= binaryThreshold ? this.opts.penValue : 0
-        }
-      }
-    } else if (sliceType === SLICE_TYPE.SAGITTAL) {
-      // Insert sagittal slice (YZ plane at X=sliceIndex)
-      for (let z = 0; z < dimZ; z++) {
-        for (let y = 0; y < dimY; y++) {
-          const srcIdx = y + z * dimY
-          const dstIdx = sliceIndex + y * dimX + z * dimX * dimY
-          this.drawBitmap[dstIdx] = slice[srcIdx] >= binaryThreshold ? this.opts.penValue : 0
-        }
-      }
-    } else {
-      throw new Error('Invalid slice type')
-    }
-  }
-
-  private smoothSlice(slice: Float32Array, width: number, height: number): void {
-    if (width < 3 || height < 3) {
-      return
-    }
-
-    const temp = new Float32Array(slice.length)
-
-    // Smooth in X direction
-    for (let y = 0; y < height; y++) {
-      for (let x = 0; x < width; x++) {
-        const idx = x + y * width
-        if (x === 0 || x === width - 1) {
-          temp[idx] = slice[idx]
-        } else {
-          temp[idx] = (slice[idx - 1] + 2 * slice[idx] + slice[idx + 1]) * 0.25
-        }
-      }
-    }
-
-    // Smooth in Y direction
-    for (let y = 0; y < height; y++) {
-      for (let x = 0; x < width; x++) {
-        const idx = x + y * width
-        if (y === 0 || y === height - 1) {
-          slice[idx] = temp[idx]
-        } else {
-          slice[idx] = (temp[idx - width] + 2 * temp[idx] + temp[idx + width]) * 0.25
-        }
-      }
-    }
-  }
-
-  private hasSliceVariability(slice: Float32Array): boolean {
-    return slice.some((value) => value > 0)
-  }
-
-  private calculateIntensityWeight(
-    intensity1: number,
-    intensity2: number,
-    targetIntensity: number,
-    intensitySigma: number
-  ): number {
-    const diff1 = Math.abs(targetIntensity - intensity1)
-    const diff2 = Math.abs(targetIntensity - intensity2)
-
-    const weight1 = Math.exp((-diff1 * diff1) / (2 * intensitySigma * intensitySigma))
-    const weight2 = Math.exp((-diff2 * diff2) / (2 * intensitySigma * intensitySigma))
-
-    const totalWeight = weight1 + weight2
-    if (totalWeight < 1e-6) {
-      return 0.5
-    }
-
-    return weight1 / totalWeight
-  }
-
-  private performGeometricInterpolation(
-    sliceLow: Float32Array,
-    sliceHigh: Float32Array,
-    z: number,
-    sliceIndexLow: number,
-    sliceIndexHigh: number,
-    interpolatedSlice: Float32Array
-  ): void {
-    const fracHigh = (z - sliceIndexLow) / (sliceIndexHigh - sliceIndexLow)
-    const fracLow = 1 - fracHigh
-
-    for (let i = 0; i < sliceLow.length; i++) {
-      interpolatedSlice[i] = sliceLow[i] * fracLow + sliceHigh[i] * fracHigh
-    }
-  }
-
-  private performIntensityGuidedInterpolation(
-    sliceLow: Float32Array,
-    sliceHigh: Float32Array,
-    z: number,
-    sliceIndexLow: number,
-    sliceIndexHigh: number,
-    interpolatedSlice: Float32Array,
-    opts: { intensityWeight: number; intensitySigma: number },
-    sliceType: SLICE_TYPE
-  ): void {
-    const intensityLow = this.extractIntensitySlice(sliceIndexLow, sliceType)
-    const intensityHigh = this.extractIntensitySlice(sliceIndexHigh, sliceType)
-    const targetIntensity = this.extractIntensitySlice(z, sliceType)
-
-    const baseFracHigh = (z - sliceIndexLow) / (sliceIndexHigh - sliceIndexLow)
-    const baseFracLow = 1 - baseFracHigh
-
-    for (let i = 0; i < sliceLow.length; i++) {
-      if (sliceLow[i] > 0 || sliceHigh[i] > 0) {
-        const intensityWeight = this.calculateIntensityWeight(
-          intensityLow[i],
-          intensityHigh[i],
-          targetIntensity[i],
-          opts.intensitySigma
-        )
-
-        const alpha = opts.intensityWeight
-        const combinedWeightLow = alpha * intensityWeight + (1 - alpha) * baseFracLow
-        const combinedWeightHigh = 1 - combinedWeightLow
-
-        interpolatedSlice[i] = sliceLow[i] * combinedWeightLow + sliceHigh[i] * combinedWeightHigh
-      } else {
-        interpolatedSlice[i] = sliceLow[i] * baseFracLow + sliceHigh[i] * baseFracHigh
-      }
-    }
+    interpolateMaskSlices(this.drawBitmap, dims, imageData, maxVal, sliceIndexLow, sliceIndexHigh, options, () =>
+      this.refreshDrawing(true)
+    )
   }
 }
